factor out some of the mutation resolvers
